Move mock logger to mock_logging.h file
